+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 [//]: # "## [unreleased] - yyyy-mm-dd"
 
+## [unreleased] - yyyy-mm-dd
+
+### Added
+
+- Filter 3σ: Filter out measurements that are more than 3σ away from the mean relative difference in comparison.
+  This filtering is done channel-wise.
+
 ## [1.2.2] - 2025-07-21
 
 ### Changed